Added game creation on lobby join
Added the creation of a new game instance whenever a user joins a lobby. This game creation includes the instantiation of a new Game with the same settings as the lobby and the respective users. The Game is then stored in the GameRepository. This change enables immediately starting a game once two users have joined the lobby.