Refactor methods and add Siren Hypermedia support in games and users actions
The commit includes a number of changes. Added support for Siren Hypermedia to games and users actions, ensuring that the responses are structured according to the Siren specification. This enhances the interoperability and navigability of the API. Also, specific methods have been renamed for better clarity and understanding. Furthermore, a typographical error in the 'winningLength' property has been corrected across various classes. For the lobbies, an additional action has been switched off temporarily with the help of comments. Additional errors related to no available lobbies have been handled. Lastly, a yet to be implemented 'TODO' has been included for future reference.